+++ packages/url_launcher/test/url_launcher_test.dart
@@ -22,15 +22,72 @@ void main() {
     await canLaunch('http://example.com/');
     expect(
       log,
-      <Matcher>[isMethodCall('canLaunch', arguments: 'http://example.com/')],
+      <Matcher>[
+        isMethodCall('canLaunch', arguments: <String, Object>{
+          'url': 'http://example.com/',
+        })
+      ],
     );
   });
 
-  test('launch', () async {
+  test('launch default behavior', () async {
     await launch('http://example.com/');
     expect(
       log,
-      <Matcher>[isMethodCall('launch', arguments: 'http://example.com/')],
+      <Matcher>[
+        isMethodCall('launch', arguments: <String, Object>{
+          'url': 'http://example.com/',
+          'useSafariVC': true,
+          'useWebView': false,
+        })
+      ],
     );
+  });
+
+  test('launch force SafariVC', () async {
+    await launch('http://example.com/', forceSafariVC: true);
+    expect(
+      log,
+      <Matcher>[
+        isMethodCall('launch', arguments: <String, Object>{
+          'url': 'http://example.com/',
+          'useSafariVC': true,
+          'useWebView': false,
+        })
+      ],
+    );
+  });
+
+  test('launch force WebView', () async {
+    await launch('http://example.com/', forceWebView: true);
+    expect(
+      log,
+      <Matcher>[
+        isMethodCall('launch', arguments: <String, Object>{
+          'url': 'http://example.com/',
+          'useSafariVC': true,
+          'useWebView': true,
+        })
+      ],
+    );
+  });
+
+  test('launch force SafariVC to false', () async {
+    await launch('http://example.com/', forceSafariVC: false);
+    expect(
+      log,
+      <Matcher>[
+        isMethodCall('launch', arguments: <String, Object>{
+          'url': 'http://example.com/',
+          'useSafariVC': false,
+          'useWebView': false,
+        })
+      ],
+    );
+  });
+
+  test('cannot launch a non-web in webview', () async {
+    expect(() async => await launch('tel:555-555-5555', forceWebView: true),
+        throwsA(const isInstanceOf<PlatformException>()));
   });
 }